% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capstone.R
\name{capstone_epub}
\alias{capstone_epub}
\title{Creates an R Markdown epub Thesis document}
\usage{
capstone_epub(...)
}
\arguments{
\item{...}{other arguments to bookdown::epub_book}
}
\value{
A ebook version of the thesis
}
\description{
This is a function called in output in the YAML of the driver Rmd file
to specify the creation of a epub version of the thesis.
}
\examples{
\dontrun{
 output: thesisdown::thesis_epub
}
}
