% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capstone.R
\name{capstone_pdf}
\alias{capstone_pdf}
\title{Creates an R Markdown PDF Capstone document}
\usage{
capstone_pdf(toc = TRUE, toc_depth = 3, highlight = "default", ...)
}
\arguments{
\item{toc}{A Boolean (TRUE or FALSE) specifying whether table of contents should be created}

\item{toc_depth}{A positive integer}

\item{highlight}{Syntax highlighting style. Supported styles include "default", "tango", "pygments", "kate", "monochrome", "espresso", "zenburn", and "haddock". Pass NULL to prevent syntax highlighting.}

\item{...}{other arguments to bookdown::pdf_book}
}
\value{
A modified \code{pdf_document} based on the Reed Senior Thesis LaTeX
  template
}
\description{
This is a function called in output in the YAML of the driver Rmd file to
specify using a UChicago Graham School MScA style LaTex and cls file.
}
\examples{
\dontrun{
 output: phoenixdown::capstone_pdf
}
}
