% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capstone.R
\name{capstone_word}
\alias{capstone_word}
\title{Creates an R Markdown Word Thesis document}
\usage{
capstone_word(...)
}
\arguments{
\item{...}{other arguments to  bookdown::word_document2}
}
\value{
A Word Document based on (hopefully soon, but not currently)
the Reed Senior Thesis Word template
}
\description{
This is a function called in output in the YAML of the driver Rmd file
to specify the creation of a Microsoft Word version of the thesis.
}
\examples{
\dontrun{
 output: thesisdown::thesis_word
}
}
