% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huskydown.R
\docType{package}
\name{huskydown}
\alias{huskydown}
\alias{huskydown-package}
\title{huskydown: A package for creating undergraduate, Masters, and PhD theses
using R Markdown}
\description{
huskydown: A package for creating undergraduate, Masters, and PhD theses
using R Markdown
}
