% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huskydown.R
\docType{package}
\name{huskydown}
\alias{huskydown}
\alias{huskydown-package}
\title{huskydown: A package for creating undergraduate, Masters, and PhD theses
using R Markdown}
\description{
huskydown: A package for creating undergraduate, Masters, and PhD theses
using R Markdown
}
\section{thesis_gitbook}{

Creates an R Markdown thesis template as a webpage
}

\section{thesis_pdf}{

Creates an R Markdown thesis template as a PDF
}

\section{thesis_word}{

Creates an R Markdown thesis template as a Microsoft Word document
}

\section{thesis_epub}{

Creates an R Markdown thesis template as an ebook
}

